Fix: revert inventory route layout
- re-introduce InventoryManagementPage and remove shared layout.
The "New warehouse" button in the shared layout would not redirect
properly. This is due to the rx.table invalidation of warehouse:list
in the page below.
