+++ apps/web-client/src/routes/inventory/inbound/[id]/+page.svelte
@@ -236,7 +236,7 @@
 	const handleSearch = async () => await goto(appPath("stock"));
 </script>
 
-<Page title={displayName} {handleCreateOutboundNote} {handleSearch}>
+<Page title={displayName} {handleCreateOutboundNote} {handleSearch} view="inbound-note">
 	<svelte:fragment slot="topbar" let:iconProps>
 		<QrCode {...iconProps} />
 		<ScannerForm
